@@ -11,7 +11,7 @@ set(PROJECT_UTILS_DIR ${PROJECT_BASE_DIR}/common)
 include_directories( ${PROJECT_BASE_DIR}/common )
 include_directories( ${PROJECT_BASE_DIR}/utils )    # needed for logentry
 
-add_definitions( -Wall -ansi -O2 -Wno-variadic-macros -std=c++11 -ggdb )
+add_definitions( -Wall -ansi -O2 -Wno-variadic-macros -std=c++17 -ggdb )
 
 add_library(common STATIC
         ${PROJECT_UTILS_DIR}/common.cpp
