@@ -11,7 +11,10 @@
 #include "daemonize.h"
 
 Camera::Server server;
-std::string logpath; 
+
+std::string log_path;             /// must set in config file
+std::string log_tmzone;           /// can set in config file, default UTC if empty
+std::string log_tostderr="true";  /// can be overridden by config file
 
 /** signal_handler ***********************************************************/
 /**
@@ -43,12 +46,13 @@ void signal_handler(int signo) {
       server.exit_cleanly();                // shutdown the server
       break;
   }
+  return;
 }
 /** signal_handler ***********************************************************/
 
 
 int  main(int argc, char **argv);           // main thread (just gets things started)
-void new_log_day( bool logstderr );         // create a new log each day
+void new_log_day();                         // create a new log each day
 void block_main(Network::TcpSocket sock);   // this thread handles requests on blocking port
 void thread_main(Network::TcpSocket sock);  // this thread handles requests on non-blocking port
 void async_main(Network::UdpSocket sock);   // this thread handles the asyncrhonous UDP message port
@@ -69,7 +73,6 @@ int main(int argc, char **argv) {
   long ret=NO_ERROR;
   std::string daemon_in;     // daemon setting read from config file
   bool start_daemon = false; // don't start as daemon unless specifically requested
-  bool logstderr = true;     // log also to stderr
 
   // capture these signals
   //
@@ -84,13 +87,16 @@ int main(int argc, char **argv) {
     if ( filename ) {
       server.config.filename = std::string( filename );
     }
+  }
+  else
 
-  } else if (argc>1) {
-      // if no "-f <filename>" then as long as there's at least one arg,
-      // assume that is the config file name.
+  // if no "-f <filename>" then as long as there's at least one arg,
+  // assume that is the config file name.
+  //
+  if (argc>1) {
     server.config.filename = std::string( argv[1] );
-
-  } else {
+  }
+  else {
     logwrite(function, "ERROR: no configuration file specified");
     server.exit_cleanly();
   }
@@ -103,43 +109,89 @@ int main(int argc, char **argv) {
   // camerad would like a few configuration keys before the daemon starts up
   //
   for (int entry=0; entry < server.config.n_entries; entry++) {
-    if (server.config.param[entry] == "LOGPATH") logpath = server.config.arg[entry];    // where to write log files
-    if (server.config.param[entry] == "TM_ZONE") zone = server.config.arg[entry];       // time zone for time stamps
+    if (server.config.param[entry] == "LOGPATH") log_path = server.config.arg[entry];   // where to write log files
     if (server.config.param[entry] == "DAEMON")  daemon_in = server.config.arg[entry];  // am I starting as a daemon or not?
+
     if (server.config.param[entry] == "LOGSTDERR") {                                    // should I log also to stderr?
       std::string stderrstr = server.config.arg[entry];
       std::transform( stderrstr.begin(), stderrstr.end(), stderrstr.begin(), ::tolower );
-      if ( stderrstr.empty() )    { continue; }          else
-      if ( stderrstr == "true"  ) { logstderr = true;  } else
-      if ( stderrstr == "false" ) { logstderr = false; } else {
-        message.str(""); message << "unrecognized value for LOGSTDERR=" << stderrstr << ": expected {True:False}";
+
+      if ( stderrstr != "true" && stderrstr != "false" ) {
+        message.str(""); message << "ERROR unknown LOGSTDERR=\"" << stderrstr << "\": expected true|false";
         logwrite( function, message.str() );
+        server.exit_cleanly();
+      }
+      else {
+        log_tostderr = stderrstr;
+        message.str(""); message << "config:" << server.config.param[entry] << "=" << server.config.arg[entry];
+        logwrite( function, message.str() );
+        server.camera.async.enqueue( message.str() );
       }
     }
+
+    // Specifies time zone for logging only, local|UTC
+    //
+    if ( server.config.param[entry] == "TM_ZONE_LOG" ) {
+      if ( server.config.arg[entry] != "UTC" && server.config.arg[entry] != "local" ) {
+        message.str(""); message << "ERROR invalid TM_ZONE_LOG=" << server.config.arg[entry] << ": expected UTC|local";
+        logwrite( function, message.str() );
+        server.exit_cleanly();
+      }
+      log_tmzone = server.config.arg[entry];
+      message.str(""); message << "config:" << server.config.param[entry] << "=" << server.config.arg[entry];
+      logwrite( function, message.str() );
+      server.camera.async.enqueue( message.str() );
+    }
+
+    // Specifies time zone for everything else, local|UTC
+    //
+    if ( server.config.param[entry] == "TM_ZONE" ) {
+      if ( server.config.arg[entry] != "UTC" && server.config.arg[entry] != "local" ) {
+        message.str(""); message << "ERROR invalid TM_ZONE=" << server.config.arg[entry] << ": expected UTC|local";
+        logwrite( function, message.str() );
+        server.exit_cleanly();
+      }
+      message.str(""); message << "TM_ZONE=" << server.config.arg[entry] << "//time zone";
+      server.systemkeys.addkey( message.str() );
+      tmzone_cfg = server.config.arg[entry];
+      message.str(""); message << "config:" << server.config.param[entry] << "=" << server.config.arg[entry];
+      logwrite( function, message.str() );
+      server.camera.async.enqueue( message.str() );
+    }
+
+    // Sets TZ environment variable (important for local time zone)
+    //
+    if (server.config.param[entry] == "TZ_ENV") {
+      setenv( "TZ", server.config.arg[entry].c_str(), 1 );
+      tzset();
+      message.str(""); message << "config:" << server.config.param[entry] << "=" << server.config.arg[entry];
+      logwrite( function, message.str() );
+      server.camera.async.enqueue( message.str() );
+    }
+
+     if (server.config.param[entry] == "DAEMON")  daemon_in = server.config.arg[entry];
+
   }
 
-  if (logpath.empty()) {
-    std::cerr << get_timestamp() << "(" << function << ") ERROR: LOGPATH not specified in configuration file\n";
+  if ( log_path.empty() ) {
+    logwrite( function, "ERROR LOGPATH not specified in configuration file" );
     server.exit_cleanly();
   }
 
-  if ( ( init_log( logpath, Camera::DAEMON_NAME, logstderr ) != 0 ) ) {  // initialize the logging system
-    std::cerr << get_timestamp() << "(" << function << ") ERROR: unable to initialize logging system\n";
+  if ( ( init_log( Camera::DAEMON_NAME, log_path, log_tostderr, log_tmzone ) != 0 ) ) {         // initialize the logging system
+    std::cerr << get_timestamp(log_tmzone) << " (" << function << ") ERROR unable to initialize logging system\n";
     server.exit_cleanly();
   }
 
-  if ( zone == "local" ) {
-    logwrite( function, "using local time zone" );
-    server.systemkeys.addkey( "TM_ZONE=local//time zone" );
-
-  } else {
-    logwrite( function, "using GMT time zone" );
-    server.systemkeys.addkey( "TM_ZONE=GMT//time zone" );
+  if ( log_tostderr.empty() ) {
+    logwrite( function, "LOGSTDERR=(empty): logs will be echoed to stderr" );
   }
 
   if ( !daemon_in.empty() && daemon_in == "yes" ) start_daemon = true;
-  else if ( !daemon_in.empty() && daemon_in == "no"  ) start_daemon = false;
-  else if ( !daemon_in.empty() ) {
+  else
+  if ( !daemon_in.empty() && daemon_in == "no"  ) start_daemon = false;
+  else
+  if ( !daemon_in.empty() ) {
     message.str(""); message << "ERROR: unrecognized argument DAEMON=" << daemon_in << ", expected { yes | no }";
     logwrite( function, message.str() );
     server.exit_cleanly();
@@ -159,7 +211,7 @@ int main(int argc, char **argv) {
 
   // log and add server build date to system keys db
   //
-  message << "this version built " << BUILD_DATE << " " << BUILD_TIME;
+  message.str(""); message << "this version built " << BUILD_DATE << " " << BUILD_TIME;
   logwrite(function, message.str());
 
   message.str(""); message << "CAMD_VER=" << BUILD_DATE << " " << BUILD_TIME << " // camerad build date";
@@ -193,9 +245,12 @@ int main(int argc, char **argv) {
   std::vector<Network::TcpSocket> socklist;          // create a vector container to hold N_THREADS TcpSocket objects
   socklist.reserve(N_THREADS);
 
-  Network::TcpSocket s(server.blkport, true, -1, 0); // instantiate TcpSocket object with blocking port
-  s.Listen();                                        // create a listening socket
-  socklist.push_back(s);                             // add it to the socklist vector
+  Network::TcpSocket sck(server.blkport, true, -1, 0); // instantiate TcpSocket object with blocking port
+  if ( sck.Listen() < 0 ) {                          // create a listening socket
+    logwrite( function, "ERROR could not create listening socket" );
+    server.exit_cleanly();
+  }
+  socklist.push_back(sck);                           // add it to the socklist vector
   std::thread(block_main, socklist[0]).detach();     // spawn a thread to handle requests on this socket
 
   // pre-thread N_THREADS-1 detached threads to handle requests on the non-blocking port
@@ -204,11 +259,14 @@ int main(int argc, char **argv) {
   for (int i=1; i<N_THREADS; i++) {                  // create N_THREADS-1 non-blocking socket objects
     if (i==1) {                                      // first one only
       Network::TcpSocket sck(server.nbport, false, CONN_TIMEOUT, i);   // instantiate TcpSocket object, non-blocking port, CONN_TIMEOUT timeout
-      sck.Listen();                                    // create a listening socket
+      if ( sck.Listen() < 0 ) {                      // create a listening socket
+        logwrite( function, "ERROR could not create listening socket" );
+        server.exit_cleanly();
+      }
       socklist.push_back(sck);
-
-    } else {                                           // subsequent socket objects are copies of the first
-      Network::TcpSocket sck = socklist[1];            // copy the first one, which has a valid listening socket
+    }
+    else {                                           // subsequent socket objects are copies of the first
+      Network::TcpSocket sck = socklist[1];          // copy the first one, which has a valid listening socket
       sck.id = i;
       socklist.push_back(sck);
     }
@@ -222,7 +280,7 @@ int main(int argc, char **argv) {
 
   // thread to start a new logbook each day
   //
-  std::thread( new_log_day, logstderr ).detach();
+  std::thread( new_log_day ).detach();
 
   for (;;) pause();                                  // main thread suspends
   return 0;
@@ -233,7 +291,7 @@ int main(int argc, char **argv) {
 /** new_log_day **************************************************************/
 /**
  * @brief      creates a new logbook each day
- * @param[in]  logstderr  true to also log to stderr
+ * @param[in]  args  optional args
  *
  * This thread is started by main and never terminates.
  * It sleeps for the number of seconds that logentry determines
@@ -243,11 +301,11 @@ int main(int argc, char **argv) {
  * is set by init_log.
  *
  */
-void new_log_day( bool logstderr ) { 
+void new_log_day() { 
   while (true) {
     std::this_thread::sleep_for( std::chrono::seconds( nextday ) );
     close_log();
-    init_log( logpath, Camera::DAEMON_NAME, logstderr );
+    init_log( Camera::DAEMON_NAME, log_path, log_tostderr, log_tmzone );
   }
 }
 /** new_log_day **************************************************************/
@@ -267,11 +325,12 @@ void new_log_day( bool logstderr ) {
  *
  */
 void block_main(Network::TcpSocket sock) {
-  while(true) {
+  while (true) {
     sock.Accept();
     doit(sock);                   // call function to do the work
     sock.Close();
   }
+  return;
 }
 /** block_main ***************************************************************/
 
@@ -301,6 +360,7 @@ void thread_main(Network::TcpSocket sock) {
     doit(sock);                // call function to do the work
     sock.Close();
   }
+  return;
 }
 /** thread_main **************************************************************/
 
@@ -342,6 +402,7 @@ void async_main(Network::UdpSocket sock) {
       return;
     }
   }
+  return;
 }
 /** async_main ***************************************************************/
 
@@ -417,7 +478,7 @@ void doit(Network::TcpSocket sock) {
     if (sbuf.empty()) {sock.Write("\n"); continue;}  // acknowledge empty command so client doesn't time out
 
     try {
-      std::size_t cmd_sep = sbuf.find_first_of(' '); // find the first space, which separates command from argument list
+      std::size_t cmd_sep = sbuf.find_first_of(" "); // find the first space, which separates command from argument list
 
       cmd = sbuf.substr(0, cmd_sep);                 // cmd is everything up until that space
 
@@ -449,14 +510,14 @@ void doit(Network::TcpSocket sock) {
      * process commands here
      */
     ret = NOTHING;
-    std::string retstring;                               // string for return the value (where needed)
+    std::string retstring="";                               // string for return the value (where needed)
 
     if (cmd=="exit") {
                     server.camera.async.enqueue("exit");    // shutdown the async message thread if running
                     server.exit_cleanly();                  // shutdown the server
                     }
     else
-    if (cmd=="config") {                         // report the config file used for camerad
+    if (cmd=="config") {                                    // report the config file used for camerad
                     std::stringstream cfg;
                     cfg << "CONFIG:" << server.config.filename;
                     server.camera.async.enqueue( cfg.str() );
@@ -557,28 +618,28 @@ void doit(Network::TcpSocket sock) {
                     }
 #ifdef ASTROCAM
     else
-    if (cmd.compare("isopen")==0) {
+    if (cmd=="isopen") {
                     ret = server.is_connected( retstring );
                     sock.Write(retstring);
                     sock.Write(" ");
                     }
     else
-    if (cmd.compare("useframes")==0) {
+    if (cmd=="useframes") {
                     ret = server.access_useframes(args);
                     if (!args.empty()) { sock.Write(args); sock.Write(" "); }
                     }
     else
-    if (cmd.compare("geometry")==0) {
+    if (cmd=="geometry") {
                     ret = server.geometry(args, retstring);
                     if (!retstring.empty()) { sock.Write(retstring); sock.Write(" "); }
                     }
     else
-    if (cmd.compare("buffer")==0) {
+    if (cmd=="buffer") {
                     ret = server.buffer(args, retstring);
                     if (!retstring.empty()) { sock.Write(retstring); sock.Write(" "); }
                     }
     else
-    if (cmd.compare("readout")==0) {
+    if (cmd=="readout") {
                     ret = server.readout(args, retstring);
                     if (!retstring.empty()) { sock.Write(retstring); sock.Write(" "); }
                     }
@@ -678,7 +739,7 @@ void doit(Network::TcpSocket sock) {
                     // Neither controller allows fractional exposure times
                     // so catch that here.
                     //
-                    if ( args.find('.') != std::string::npos ) {
+                    if ( args.find(".") != std::string::npos ) {
                       ret = ERROR;
                       logwrite(function, "ERROR: fractional exposure times not allowed");
                       // empty the args string so that a call to exptime returns the current exptime
@@ -746,6 +807,7 @@ void doit(Network::TcpSocket sock) {
   }
 
   sock.Close();
+  return;
 }
 /** doit *********************************************************************/
 
