@@ -5,8 +5,7 @@
  *
  */
 
-#ifndef COMMON_H
-#define COMMON_H
+#pragma once
 
 #include <sstream>
 #include <queue>
@@ -40,6 +39,93 @@ namespace Common {
       std::string get_keytype(std::string keyvalue);         /// return type of keyword based on value
       long listkeys();                                       /// list FITS keys in the internal database
       long addkey(std::string arg);                          /// add FITS key to the internal database
+      long delkey(std::string arg);                          /// delete FITS key from the internal database
+      void erasedb() { this->keydb.clear(); };               /// erase the entire contents of the internal database
+
+      /***** Common::FitsKeys::addkey *****************************************/
+      /**
+       * @brief      template function adds FITS keyword to internal database
+       * @details    no parsing is done here
+       * @param[in]  keyword  keyword name <string>
+       * @param[in]  value    value for keyword, any type <T>
+       * @param[in]  comment  comment <string>
+       * @param[in]  prec     decimal places of precision
+       * @return     ERROR or NO_ERROR
+       *
+       * This function is overloaded.
+       *
+       * This version adds the keyword=value//comment directly into the
+       * database with no parsing. The database stores only strings so
+       * the type variable preserves the type and informs the FITS writer.
+       *
+       * The template input "tval" here allows passing in any type of value,
+       * and the type will be set by the actual input type. However, if
+       * the input type is a string then an attempt will be made to infer the
+       * type from the string (this allows passing in a string "12.345" and
+       * have it be converted to a double.
+       *
+       * The prec argument sets the displayed precision of double or float
+       * types if passed in as a double or float. Default is 8 if not supplied.
+       *
+       * The overloading for type bool is because for the template class,
+       * since T can be evaluated as string, it can't be treated as bool,
+       * so this is the easiest way to extract the bool value.
+       *
+       */
+      inline long addkey( const std::string &key, bool bval, const std::string &comment ) {
+        return addkey( key, (bval?"T":"F"), comment, 0 );
+      }
+      template <class T> long addkey( const std::string &key, T tval, const std::string &comment ) {
+        return addkey( key, tval, comment, 8 );
+      }
+      template <class T> long addkey( const std::string &key, T tval, const std::string &comment, int prec ) {
+        std::stringstream val;
+        std::string type;
+
+        // "convert" the type <T> value into a string via the appropriate stream
+        // and set the type string
+        //
+        if ( std::is_same<T, double>::value ) {
+          val << std::fixed << std::setprecision( prec ) << tval;
+          type = "DOUBLE";
+        }
+        else
+        if ( std::is_same<T, float>::value ) {
+          val << std::fixed << std::setprecision( prec ) << tval;
+          type = "FLOAT";
+        }
+        else
+        if ( std::is_same<T, int>::value ) {
+          val << tval;
+          type = "INT";
+        }
+        else
+        if ( std::is_same<T, long>::value ) {
+          val << tval;
+          type = "LONG";
+        }
+        else {
+          val << tval;
+          type = this->get_keytype( val.str() );  // try to infer the type from any string
+        }
+
+        // insert new entry into the database
+        //
+        this->keydb[key].keyword    = key;
+        this->keydb[key].keytype    = type;
+        this->keydb[key].keyvalue   = val.str();
+        this->keydb[key].keycomment = comment;
+
+#ifdef LOGLEVEL_DEBUG
+        std::string function = "Common::FitsKeys::addkey";
+        std::stringstream message;
+        message << "[DEBUG] added key: " << key << "=" << tval << " (" << this->keydb[key].keytype << ") // " << comment;
+        logwrite( function, message.str() );
+#endif
+        return( NO_ERROR );
+      }
+      /***** Common::FitsKeys::addkey *****************************************/
+
 
       typedef struct {                                       /// structure of FITS keyword internal database
         std::string keyword;
@@ -119,4 +205,3 @@ namespace Common {
   /**************** Common::Queue *********************************************/
 
 }
-#endif
