@@ -1,7 +1,6 @@
 /**
  * @file    logentry.cpp
  * @brief   logentry functions
- * @details 
  * @author  David Hale <dhale@astro.caltech.edu>
  *
  * Write time-stamped entries to log file using logwrite(function, message)
@@ -12,34 +11,38 @@
 
 #include "logentry.h"
 
-std::mutex loglock;
+std::mutex loglock;            /// mutex to protect from multiple access
 std::ofstream filestream;      /// IO stream class
 unsigned int nextday = 86410;  /// number of seconds until a new day
-bool to_stderr = true;         /// write to stderr
+bool to_stderr = true;         /// write to stderr by default
+std::string tmzone_log;        /// optional time zone for logging
 
 
-/** init_log *****************************************************************/
+/***** init_log ***************************************************************/
 /**
- * @fn         init_log
  * @brief      initializes the logging
- * @param[in]  logpath string is the path for log file
- * @param[in]  name string is the name of the log file in logpath
- * @param[in]  stderr_in  true to log also to stderr
+ * @param[in]  name       name of the log file in logpath
+ * @param[in]  logpath    path for log file
+ * @param[in]  logstderr  should logwrite also print to stderr?
+ * @param[in]  logtmzone  optional time zone for logging
  * @return     0 on success, 1 on error
  *
  * Opens an ofstream to the specified logfile, "logpath/name_YYYYMMDD.log"
  * where logpath and name are passed in as parameters.
  *
  */
-long init_log( std::string logpath, std::string name, bool stderr_in ) {
-  std::string function = "init_log";
+long init_log( std::string name, std::string logpath, std::string logstderr, std::string logtmzone ) {
+  const std::string function = "init_log";
   std::stringstream filename;
   std::stringstream message;
   int year, mon, mday, hour, min, sec, usec;
   long error = 0;
-  to_stderr = stderr_in;
 
-  if ( ( error = get_time( year, mon, mday, hour, min, sec, usec ) ) ) return error;
+  to_stderr = ( logstderr=="false" ? false : true );  // should logwrite also print to stderr?
+
+  tmzone_log = logtmzone;
+
+  if ( ( error = get_time( tmzone_log, year, mon, mday, hour, min, sec, usec ) ) ) return error;
 
   // assemble log file name from the passed-in arguments and current date
   //
@@ -57,48 +60,87 @@ long init_log( std::string logpath, std::string name, bool stderr_in ) {
   try {
     filestream.open(filename.str(), std::ios_base::app);
   }
-  catch(...) {
-    message << "ERROR: opening log file " << filename.str() << ": " << std::strerror(errno);
+  catch ( const std::filesystem::filesystem_error& e ) {
+    message.str(""); message << "ERROR " << filename.str() << ": " << e.what() << ": " << e.code().value();
     logwrite(function, message.str());
     return 1;
   }
+  catch(...) {
+    message.str(""); message << "ERROR: opening log file " << filename.str() << ": " << std::strerror(errno);
+    logwrite(function, message.str());
+    return 1;
+  }
+
+  // If I am the owner then make sure the permissions are set correctly.
+  // Only the owner can change permissions. Even if I have write access,
+  // if I'm not the owner then I can't change the permissions.
+  //
+  if ( is_owner( filename.str() ) ) {
+    try {
+      // remove all permissions
+      //
+      std::filesystem::permissions( filename.str(),
+                                    std::filesystem::perms::all,
+                                    std::filesystem::perm_options::remove
+                                  );
+
+      // add back permissions rw-rw-r-- (664)
+      //
+      std::filesystem::permissions( filename.str(),
+                                    std::filesystem::perms::owner_read  |
+                                    std::filesystem::perms::owner_write |
+                                    std::filesystem::perms::group_read  |
+                                    std::filesystem::perms::group_write |
+                                    std::filesystem::perms::others_read,
+                                    std::filesystem::perm_options::add
+                                  );
+    }
+    catch ( const std::filesystem::filesystem_error& e ) {
+      message.str(""); message << "ERROR " << filename.str() << ": " << e.what() << ": " << e.code().value();
+      logwrite(function, message.str());
+      return 1;
+    }
+  }
+
+  // If I do not have write permission then that is a fatal error
+  //
+  if ( ! has_write_permission( filename.str() ) ) {
+    message.str(""); message << "ERROR: no write permission for log file " << filename.str();
+    logwrite(function, message.str());
+    return 1;
+  }
+
   if ( ! filestream.is_open() ) {
-    message << "ERROR: log file " << filename.str() << " not open";
+    message.str(""); message << "ERROR: log file " << filename.str() << " not open";
     logwrite(function, message.str());
     return 1;
   }
 
   return 0;
 }
-/** init_log *****************************************************************/
+/***** init_log ***************************************************************/
 
 
-/** close_log ****************************************************************/
+/***** close_log **************************************************************/
 /**
- * @fn         close_log
  * @brief      closes the logfile stream
- * @param[in]  none
- * @return     none
- *
- * Call this from the main class deconstructor to clean up the log file.
+ * @details    Call this from the main class deconstructor to clean up the log file.
  *
  */
 void close_log() {
-  if (filestream.is_open() == true) {
+  if ( filestream.is_open() ) {
     filestream.flush();
     filestream.close();
   }
 }
-/** close_log ****************************************************************/
+/***** close_log **************************************************************/
 
 
-/** logwrite *****************************************************************/
+/***** logwrite ***************************************************************/
 /**
- * @fn         logwrite
  * @brief      create a log file entry
- * @param[in]  function
- * @param[in]  message
- * @return     none
+ * @param[in]  function  string containing the Namespace::Class::function
+ * @param[in]  message   string to log
  *
  * Create a time-stamped entry in the log file in the form of:
  * YYYY-MM-DDTHH:MM:SS.ssssss (function) message\n
@@ -107,11 +149,12 @@ void close_log() {
  * log filestream isn't open.
  *
  */
-void logwrite(std::string function, std::string message) {
+void logwrite( const std::string &function, const std::string &message ) {
   std::stringstream logmsg;
-  std::string timestamp = get_timestamp();       // get the current time (defined in utilities.h)
 
   std::lock_guard<std::mutex> lock(loglock);     // lock mutex to protect from multiple access
+
+  std::string timestamp = get_timestamp(tmzone_log);  // get the current time (defined in utilities.h)
 
   logmsg << timestamp << "  (" << function << ") " << message << "\n";
 
@@ -121,5 +164,5 @@ void logwrite(std::string function, std::string message) {
   }
   if ( to_stderr ) std::cerr << logmsg.str();    // send to standard error if requested
 }
-/** logwrite *****************************************************************/
+/***** logwrite ***************************************************************/
 
